services/base-template/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.32.1
pydantic==2.9.2

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Observability
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-instrumentation-sqlalchemy==0.49b2
opentelemetry-exporter-otlp==1.28.2
prometheus-client==0.21.0
structlog==24.4.0

# Utils
python-json-logger==2.0.7
httpx==0.28.0
tenacity==9.0.0
